requirements.txt
alembic==1.0.0
amqp==2.3.2
aniso8601==3.0.2
billiard==3.5.0.4
bleach==3.0.1
blinker==1.4
Bootstrap-Flask==1.0.8
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
comtypes==1.1.7
dnspython==1.15.0
dominate==2.3.4
eventlet==0.24.1
Faker==0.9.1
Flask==1.0.2
Flask-Avatars==0.2.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-CKEditor==0.4.2
Flask-DebugToolbar==0.10.1
Flask-Dropzone==1.5.3
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flask-whooshee==0.6.0
Flask-WTF==0.14.2
greenlet==0.4.15
html5lib==1.0.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
monotonic==1.5
Pillow==5.3.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
pywifi==1.1.10
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.12
text-unidecode==1.2
urllib3==1.23
vine==1.1.4
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
